Changed file name and added modifications to pass all tests in check_estimator
